Allow different sites to have different basis frame for projections (#377)
