+++ examples/sym_wann_2/example.py
@@ -38,7 +38,7 @@ t222 = time()
 # proj2 = Projection( (np.array([0,0,0]),"t2g" , spacegroup=spacegroup)
 symmetrizer.set_spacegroup(spacegroup)
 t223 = time()
-symmetrizer.set_D_wann_from_projections( projections_obj=[proj1,proj2])
+symmetrizer.set_D_wann_from_projections( projections=[proj1,proj2])
 t22 = time()
 system_2.symmetrize2(symmetrizer)
 t23 = time()
