+++ examples/wannierise+sitesym_Fe/wannierise_Fe.py
@@ -23,7 +23,7 @@ projection_t2g = Projection(orbital='t2g', position_num=[0,0,0], spacegroup=spac
 # projections_set = ProjectionsSet(projections=[projection_s, projection_p, projection_d])
 projections_set = ProjectionsSet(projections=[projection_sp3d2, projection_t2g])
 
-symmetrizer.set_D_wann_from_projections(projections_obj=projections_set)
+symmetrizer.set_D_wann_from_projections(projections_set)
 amn = wb.w90files.amn_from_bandstructure(bandstructure, projections_set=projections_set)
 w90data.set_symmetrizer(symmetrizer)
 w90data.set_file("amn", amn)
