+++ tests/test_find_projections.py
@@ -5,8 +5,8 @@ import numpy as np
 import sympy
 from wannierberri.symmetry.sawf import SymmetrizerSAWF
 
-from wannierberri.wannierise.projections_searcher import EBRsearcher
-from wannierberri.wannierise.projections import Projection, ProjectionsSet
+from wannierberri.symmetry.projections_searcher import EBRsearcher
+from wannierberri.symmetry.projections import Projection, ProjectionsSet
 
 from .common import OUTPUT_DIR, ROOT_DIR
 
@@ -16,7 +16,7 @@ def test_find_projections_diamond():
     bandstructure = BandStructure(prefix=data_dir + "/di", code="espresso",
                             Ecut=100, include_TR=False)
     spacegroup = bandstructure.spacegroup
-    # spacegroup.show()
+    spacegroup.show()
 
     symmetrizer = SymmetrizerSAWF().from_irrep(bandstructure)
     symmetrizer.to_npz(OUTPUT_DIR + "/diamond-only-bands.sawf.npz")
