+++ wannierberri/system/system_tb.py
@@ -92,7 +92,7 @@ class System_tb(System_R):
                 wannier_centers_cart = np.diagonal(AA_R[:, :, self.iR0, :], axis1=0, axis2=1).T.copy()
             if convention_II_to_I:
                 # convert to convention I
-                print(f"convention_II_to_I = {convention_II_to_I} wannier_centers_cart = \n{wannier_centers_cart}\n num_wann = {self.num_wann}, A.shape = {AA_R.shape}")
+                # print(f"convention_II_to_I = {convention_II_to_I} wannier_centers_cart = \n{wannier_centers_cart}\n num_wann = {self.num_wann}, A.shape = {AA_R.shape}")
                 AA_R[np.arange(self.num_wann), np.arange(self.num_wann), self.iR0, :] -= wannier_centers_cart
             self.set_R_mat('AA', AA_R)
         elif wannier_centers_cart is None:
