+++ wannierberri/w90files/__init__.py
@@ -3,10 +3,8 @@ from .win import WIN
 from .eig import EIG
 from .mmn import MMN
 from .amn import AMN, amn_from_bandstructure
-from .dmn import DMN
 from .xxu import UIU, UHU, SIU, SHU
 from .spn import SPN
-from .dmn import DMN
 
 from .chk import CheckPoint, CheckPoint_bare
 from .w90data import Wannier90data
