+++ wannierberri/wannierise/__init__.py
@@ -1,2 +1,1 @@
 from .wannierise import wannierise
-from .projections_searcher import EBRsearcher
