+++ wannierberri/workflow/workflow.py
@@ -528,10 +528,9 @@ class WorkflowQE:
         if enforce or not self.flags.check('dmn'):
             dmn_new = DMN(empty=True)
             dmn_new.from_irrep(bandstructure)
-            dmn_new.set_D_wann_from_projections(projections_obj=self.projections, spinor=self.spinor)
+            dmn_new.set_D_wann_from_projections(self.projections)
             dmn_new.to_w90_file(self.prefix)
             self.flags.on('dmn')
-
 
 
     def calc_bands_wannier_w90(self, kdensity=1000, enforce=False):
